new function testring-cli run
